requirements.txt
python-telegram-bot
paho-mqtt
meshtastic
pycryptodome 